requirements.txt
python-dotenv==1.0.0
selenium==4.15.2
playwright==1.40.0
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2
plotly==5.18.0
dash==2.14.1
requests==2.31.0
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1 